Control sesion

+++ main.py
@@ -2,6 +2,7 @@ import telebot
 import json
 from configs.secrets import token
 from libs.comandos import Commands
+from libs.mensajes import msg_start_with, msg_add_product_end, msg_add_product_error, msg_add_product_error_slash, msg_unauthorized
 
 
 bot = telebot.TeleBot(token)
@@ -12,17 +13,23 @@ shopping_list = 'data/shopping_list.json'
 
 @bot.message_handler(commands=['help', 'ayuda'])
 def cmd_handler(message):
-    cmd.help(message)
+        cmd.help(message)
 
 
 @bot.message_handler(commands=['list', 'lista'])
 def cmd_handler(message):
-    cmd.show_list(message)
+    if check_user_id(message):
+        cmd.show_list(message)
+    else:
+        bot.send_message(message.chat.id, msg_unauthorized)
 
 
 @bot.message_handler(commands=['clear', 'borrar'])
 def cmd_handler(message):
-    cmd.delete_list(message)
+    if check_user_id(message):
+        cmd.delete_list(message)
+    else:
+        bot.send_message(message.chat.id, msg_unauthorized)
 
 # @bot.message_handler(commands=['add', 'agregar'])
 # def cmd_handler(message):
@@ -30,14 +37,14 @@ def cmd_handler(message):
 
 
 @bot.message_handler(commands=['add', 'agregar'])
-def bot_mensajes_texto(message):
+def add_product_to_list(message):
     if message.text.startswith("/"):
-        bot.send_message(message.chat.id, 'Agrega productos a la lista. Una vez finalizado usa el comando /end para salir del modo de agregación.')
+        bot.send_message(message.chat.id, msg_start_with)
 
         # Función para manejar el siguiente mensaje
         def handle_next_step(next_message):
             if next_message.text == '/end':
-                bot.send_message(next_message.chat.id, 'Modo de agregación finalizado.')
+                bot.send_message(next_message.chat.id, msg_add_product_end)
             elif not next_message.text.startswith("/"):
                 with open(shopping_list, "r") as archivo:
                     datos = json.load(archivo)
@@ -50,21 +57,26 @@ def bot_mensajes_texto(message):
                     json.dump(datos, archivo)
 
                 # Responder con un emoji
-                emoji_producto_agregado = "\u2705"
-                bot.send_message(next_message.chat.id, emoji_producto_agregado)
+                emoji_product_added = "\u2705"
+                bot.send_message(next_message.chat.id, emoji_product_added)
 
                 # Espera al siguiente mensaje
                 bot.register_next_step_handler(next_message, handle_next_step)
             else:
-                bot.send_message(next_message.chat.id, 'El producto no puede empezar con "/".')
+                bot.send_message(next_message.chat.id, msg_add_product_error_slash)
 
         # Espera al siguiente mensaje
         bot.register_next_step_handler(message, handle_next_step)
     else:
-        bot.send_message(message.chat.id, 'Comando inválido. Por favor, usa el comando /add o /agregar para iniciar el modo de agregación.')
+        bot.send_message(message.chat.id, msg_add_product_error)
 
 
 
+def check_user_id(message):
+    if message.from_user and message.from_user.id == 504727906:
+        return True
+    else:
+        return False
 
 
 if __name__ == '__main__':
